fix Bug

+++ backend/gonlink/src/main/java/online/gonlink/service/impl/UrlShortenerServiceImpl.java
@@ -132,6 +132,12 @@ public class UrlShortenerServiceImpl implements UrlShortenerService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public RemoveUrlResponse removeByShortCode(RemoveUrlRequest request) {
+        GeneralTraffic generalTraffic = trafficService.searchGeneralTrafficByShortCode(request.getShortCode());
+        if(Objects.isNull(generalTraffic))
+            throw new ResourceException(ExceptionEnum.NOT_FOUND_SHORT_CODE, null);
+
+
+
         shortUrlRep.deleteById(request.getShortCode());
         trafficService.deletesTraffic(request);
         return RemoveUrlResponse.newBuilder()
@@ -151,7 +157,7 @@ public class UrlShortenerServiceImpl implements UrlShortenerService {
 
         /* request.getTimeExpired() = 'null' means no timeExpired will be used */
         if(request.getTimeExpired().equals("null")){
-            shortUrl.setTimeExpired("");
+            shortUrl.setTimeExpired(null);
         } else{
             if (!request.getTimeExpired().equals("")){
                 ZonedDateTime zonedDateTime = ZonedDateTime.parse(request.getTimeExpired()).withZoneSameInstant(ZoneId.of(request.getZoneId()));
@@ -167,8 +173,8 @@ public class UrlShortenerServiceImpl implements UrlShortenerService {
                 .setOriginalUrl(shortUrlUpdated.getOriginalUrl())
                 .setAlias(shortUrlUpdated.getAlias())
                 .setDesc(shortUrlUpdated.getDesc())
-                .setPassword(Objects.isNull(shortUrlUpdated.getPassword())?"":shortUrlUpdated.getPassword())
-                .setTimeExpired(Objects.isNull(shortUrlUpdated.getTimeExpired())?"":shortUrlUpdated.getTimeExpired().toString())
+                .setPassword(Objects.isNull(shortUrlUpdated.getPassword())?null:shortUrlUpdated.getPassword())
+                .setTimeExpired(Objects.isNull(shortUrlUpdated.getTimeExpired())?null:shortUrlUpdated.getTimeExpired().toString())
                 .setMaxUsage(shortUrlUpdated.getMaxUsage())
                 .setActive(shortUrlUpdated.isActive())
                 .build();
